Time: 33 ms (26.88%) | Memory: 63.1 MB (58.60%) - LeetSync

+++ 1335-maximum-candies-allocated-to-k-children/maximum-candies-allocated-to-k-children.java
@@ -1,8 +1,8 @@
 class Solution {
     public int maximumCandies(int[] candies, long k) {
-        long l = 1, r = 10_000_000 ;
+        int l = 1, r = 10_000_000 ;
         while(l <= r){
-            long md = l + (r - l)/2 ;
+            int md = l + (r - l)/2 ;
             long cnt = 0 ;
             for(int i : candies){
                 cnt += (i/md);
@@ -11,6 +11,6 @@ class Solution {
                 l = md + 1;
             }else r = md - 1;
         }
-        return (int)r ;
+        return r ;
     }
 }
